improve context javadoc, changing method signatures to be safer, package change
