+++ src/main/java/com/github/syr0ws/craftventory/internal/inventory/SimpleCraftVentory.java
@@ -23,7 +23,7 @@ import com.github.syr0ws.craftventory.internal.inventory.data.SimpleDataStore;
 import com.github.syr0ws.craftventory.internal.inventory.hook.SimpleHookManager;
 import com.github.syr0ws.craftventory.internal.inventory.item.SimpleInventoryItem;
 import com.github.syr0ws.craftventory.internal.inventory.pagination.SimplePaginationManager;
-import com.github.syr0ws.craftventory.internal.util.SimpleContext;
+import com.github.syr0ws.craftventory.common.util.SimpleContext;
 import org.bukkit.Bukkit;
 import org.bukkit.event.inventory.InventoryType;
 import org.bukkit.inventory.Inventory;
