+++ src/main/java/com/github/syr0ws/craftventory/internal/inventory/SimpleInventoryViewManager.java
@@ -5,7 +5,7 @@ import com.github.syr0ws.craftventory.api.inventory.data.DataStore;
 import com.github.syr0ws.craftventory.api.inventory.InventoryViewManager;
 import com.github.syr0ws.craftventory.api.util.Context;
 import com.github.syr0ws.craftventory.internal.inventory.data.SimpleDataStore;
-import com.github.syr0ws.craftventory.internal.util.SimpleContext;
+import com.github.syr0ws.craftventory.common.util.SimpleContext;
 
 import java.util.ArrayList;
 import java.util.List;
